requirements.txt
streamlit
supabase
requests
python-dotenv
pandas
